pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "oak-ci"
dynamic = ["version"]
description = "CLI toolkit for AI-powered development workflows - spec-driven development, RFC/ARC management, and shared conventions across AI coding agents"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
license-files = ["LICENSE"]
authors = [
    {name = "Chris Kirby"}
]
keywords = ["cli", "ai", "rfc", "development-workflow", "ai-agents", "claude", "copilot", "cursor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Software Development :: Quality Assurance",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]

dependencies = [
    # Core CLI dependencies
    "typer>=0.12.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "readchar>=4.0.0",
    "platformdirs>=4.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.26",

    # Codebase Intelligence core deps (no parsers - parsers installed via oak languages)
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "chromadb>=0.5.0",
    "tree-sitter>=0.23.0",  # Base library only
    "watchdog>=4.0.0",
    "aiofiles>=23.0.0",
    "mcp>=1.0.0",
    "croniter>=2.0.0",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=26.0.0",
    "types-PyYAML>=6.0.0",
]

# Individual language parsers (installed via oak languages add)
parser-python = ["tree-sitter-python>=0.23.0"]
parser-javascript = ["tree-sitter-javascript>=0.23.0"]
parser-typescript = ["tree-sitter-typescript>=0.23.0"]
parser-java = ["tree-sitter-java>=0.23.0"]
parser-csharp = ["tree-sitter-c-sharp>=0.23.0"]
parser-go = ["tree-sitter-go>=0.23.0"]
parser-rust = ["tree-sitter-rust>=0.23.0"]
parser-c = ["tree-sitter-c>=0.23.0"]
parser-cpp = ["tree-sitter-cpp>=0.23.0"]
parser-ruby = ["tree-sitter-ruby>=0.23.0"]
parser-php = ["tree-sitter-php>=0.23.0"]
parser-kotlin = ["tree-sitter-kotlin>=0.23.0"]
parser-scala = ["tree-sitter-scala>=0.23.0"]

# Convenience groups for common use cases
parsers-common = [
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
]

parsers-all = [
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-c-sharp>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-c>=0.23.0",
    "tree-sitter-cpp>=0.23.0",
    "tree-sitter-ruby>=0.23.0",
    "tree-sitter-php>=0.23.0",
    "tree-sitter-kotlin>=0.23.0",
    "tree-sitter-scala>=0.23.0",
]

[project.scripts]
oak = "open_agent_kit.cli:cli_main"

[project.urls]
Homepage = "https://github.com/goondocks-co/open-agent-kit"
Documentation = "https://github.com/goondocks-co/open-agent-kit/blob/main/README.md"
Repository = "https://github.com/goondocks-co/open-agent-kit"
Issues = "https://github.com/goondocks-co/open-agent-kit/issues"
Changelog = "https://github.com/goondocks-co/open-agent-kit/releases"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/open_agent_kit/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/open_agent_kit"]
# Include non-Python files in wheel (config, templates, manifests, static assets)
artifacts = [
    "src/open_agent_kit/agents/**/*.yaml",
    "src/open_agent_kit/features/**/*.yaml",
    "src/open_agent_kit/features/**/*.md",
    "src/open_agent_kit/features/**/*.json",
    "src/open_agent_kit/features/**/*.sh",
    # CI daemon UI static files
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.html",
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.css",
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.js",
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.png",
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.svg",
    "src/open_agent_kit/features/codebase_intelligence/daemon/static/**/*.ico",
]
# Exclude node_modules and other development artifacts
exclude = [
    "**/node_modules/**",
    "**/ui/src/**",
    "**/.vite/**",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/templates",
    "/scripts",
    "/install.sh",
    "/install.ps1",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=open_agent_kit --cov-report=term-missing --cov-report=html"
filterwarnings = [
    # Ignore ResourceWarning from mocked SQLite connections in tests
    "ignore:unclosed.*sqlite3.Connection:ResourceWarning",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
files = ["src/open_agent_kit"]

# Optional dependencies without type stubs
[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true

# Optional codebase-intelligence feature dependencies
# These are only installed with 'make setup-full' or 'oak init --enable-ci'
[[tool.mypy.overrides]]
module = [
    "fastapi",
    "fastapi.*",
    "chromadb",
    "chromadb.*",
    "tree_sitter",
    "tree_sitter.*",
    "watchdog",
    "watchdog.*",
    "aiofiles",
    "aiofiles.*",
    "uvicorn",
    "uvicorn.*",
    "claude_agent_sdk",
    "claude_agent_sdk.*",
    "croniter",
    "croniter.*",
    "tomli_w",
]
ignore_missing_imports = true

# Codebase Intelligence daemon files have type inference issues when
# optional deps (fastapi, etc.) are not installed - relax type checking
[[tool.mypy.overrides]]
module = "open_agent_kit.features.codebase_intelligence.daemon.*"
disable_error_code = ["no-any-return"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["src/open_agent_kit/_version.py"]  # Auto-generated by hatch-vcs

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "UP042",  # str+Enum → StrEnum migration: RFC-001 §8 Phase 3 #31 — remove this ignore after migration
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused
"src/open_agent_kit/commands/*.py" = ["B904"]  # typer.Exit doesn't need exception chaining
"src/open_agent_kit/models/agent.py" = ["B904"]  # re-raising with custom message is intentional

[tool.black]
line-length = 100
target-version = ["py312", "py313"]
include = '\.pyi?$'
